Use unsigned int in fifof/label as output type
The unsigned short proved to be not sufficient for the task.

+++ mia/core/labelmap.hh
@@ -33,7 +33,7 @@ NS_MIA_BEGIN
    \ingroup misc
    \brief A simple class to add loadind and storeing to a map of labels. 
 */
-class EXPORT_CORE CLabelMap: public std::map<unsigned short,  unsigned short> {
+class EXPORT_CORE CLabelMap: public std::map<unsigned int,  unsigned int> {
 public: 
 	CLabelMap() = default;
